Update tech crew's service and controller, make delete service method throw a not found exc. and catch it in the controller

+++ src/main/java/com/alpdogan/CompanyManagementSystem/service/TechCrewService.java
@@ -71,14 +71,23 @@ public class TechCrewService {
 
     }
 
-    public String deleteTechCrewById(Integer techCrewId) {
+    public String deleteTechCrewById(Integer techCrewId) throws Exception {
 
         Optional<TechCrew> techCrewOptional = techCrewRepository.findById(techCrewId);
-        TechCrew techCrew = techCrewOptional.get();
 
-        techCrewRepository.delete(techCrew);
+        if (techCrewOptional.isPresent()) {
 
-        return "Tech Crew Deleted.";
+            TechCrew techCrew = techCrewOptional.get();
+
+            techCrewRepository.delete(techCrew);
+
+            return "Tech Crew Deleted.";
+
+        }else {
+
+            throw new Exception("Tech Crew Not Found.");
+
+        }
 
     }
